#42 Refactor client directory structure
