+++ src/client/src/modules/StatsModal.tsx
@@ -1,7 +1,7 @@
 import React, { useEffect, useState } from "react";
 import queryString from "query-string";
 import Tooltip from "@mui/material/Tooltip";
-import { countScores } from "../../utils/stats.utils";
+import { countScores } from "../utils/stats.utils";
 
 interface StatsBarHeightsState {
   [key: number]: number;
@@ -45,7 +45,8 @@ const StatsModal: React.FC<StatsModalProps> = ({
           <Tooltip title={scores[1]} placement="bottom" arrow>
             <div
               className={`bg-green w-4`}
-              style={{ height: `${statsBarHeights[1]}px` }}></div>
+              style={{ height: `${statsBarHeights[1]}px` }}
+            ></div>
           </Tooltip>
         </div>
         <div className="relative ml-6">
@@ -55,7 +56,8 @@ const StatsModal: React.FC<StatsModalProps> = ({
           <Tooltip title={scores[2]} placement="bottom" arrow>
             <div
               className={`bg-green w-4`}
-              style={{ height: `${statsBarHeights[2]}px` }}></div>
+              style={{ height: `${statsBarHeights[2]}px` }}
+            ></div>
           </Tooltip>
         </div>
         <div className="relative ml-6">
@@ -65,7 +67,8 @@ const StatsModal: React.FC<StatsModalProps> = ({
           <Tooltip title={scores[3]} placement="bottom" arrow>
             <div
               className={`bg-green w-4`}
-              style={{ height: `${statsBarHeights[3]}px` }}></div>
+              style={{ height: `${statsBarHeights[3]}px` }}
+            ></div>
           </Tooltip>
         </div>
         <div className="relative ml-6">
@@ -75,7 +78,8 @@ const StatsModal: React.FC<StatsModalProps> = ({
           <Tooltip title={scores[4]} placement="bottom" arrow>
             <div
               className={`bg-green w-4`}
-              style={{ height: `${statsBarHeights[4]}px` }}></div>
+              style={{ height: `${statsBarHeights[4]}px` }}
+            ></div>
           </Tooltip>
         </div>
         <div className="relative ml-6">
@@ -85,7 +89,8 @@ const StatsModal: React.FC<StatsModalProps> = ({
           <Tooltip title={scores[5]} placement="bottom" arrow>
             <div
               className={`bg-green w-4`}
-              style={{ height: `${statsBarHeights[5]}px` }}></div>
+              style={{ height: `${statsBarHeights[5]}px` }}
+            ></div>
           </Tooltip>
         </div>
         <div className="relative ml-6">
@@ -95,7 +100,8 @@ const StatsModal: React.FC<StatsModalProps> = ({
           <Tooltip title={scores[6]} placement="bottom" arrow>
             <div
               className={`bg-green w-4`}
-              style={{ height: `${statsBarHeights[6]}px` }}></div>
+              style={{ height: `${statsBarHeights[6]}px` }}
+            ></div>
           </Tooltip>
         </div>
         <div className="relative ml-6">
@@ -105,7 +111,8 @@ const StatsModal: React.FC<StatsModalProps> = ({
           <Tooltip title={scores[0]} placement="bottom" arrow>
             <div
               className={`bg-red w-4`}
-              style={{ height: `${statsBarHeights[0]}px` }}></div>
+              style={{ height: `${statsBarHeights[0]}px` }}
+            ></div>
           </Tooltip>
         </div>
       </div>
