+++ src/client/src/pages/BaseGame.tsx
@@ -1,12 +1,9 @@
 import React, { useEffect, useState } from "react";
-import { useLocation } from "react-router-dom";
-import queryString from "query-string";
 import Modal from "react-modal";
 
 import NavBar from "../modules/Navbar";
 import Game from "../modules/Game";
 import GameConclusion from "../modules/GameConclusion";
-import PlaylistSearch from "../modules/PlaylistSearch";
 import HelpModal from "../modules/HelpModal";
 import StatsModal from "../modules/StatsModal";
 import UserAccountModal from "../modules/UserAccountModal";
@@ -15,18 +12,18 @@ import Loader from "../modules/Loader";
 import {
   calculateBarHeights,
   calculateStatsBottom,
-} from "../../utils/stats.utils";
+} from "../utils/stats.utils";
 
-import GameResult from "../interfaces/GameResult";
-import SavedGameData from "../interfaces/SavedGameData";
-import TrackGuessFormat from "../interfaces/TrackGuessFormat";
-import TrackFormat from "../interfaces/TrackFormat";
+import GameResult from "../types/GameResult";
+import SavedGameData from "../types/SavedGameData";
+import TrackGuessFormat from "../types/TrackGuessFormat";
+import TrackFormat from "../types/TrackFormat";
 
 interface StatsBarHeightsState {
   [key: number]: number;
 }
 
-const CustomGame: React.FC<{ apiOrigin: string }> = ({ apiOrigin }) => {
+const BaseGame: React.FC<{ apiOrigin: string }> = ({ apiOrigin }) => {
   const [song, setSong] = useState<string>("");
   const [artists, setArtists] = useState<string[]>([]);
   const [id, setId] = useState<number>(0);
@@ -37,7 +34,6 @@ const CustomGame: React.FC<{ apiOrigin: string }> = ({ apiOrigin }) => {
   const [userGuesses, setUserGuesses] = useState<TrackGuessFormat[]>([]);
 
   const [gameFinished, setGameFinished] = useState<boolean>(false);
-  const [validPlaylist, setvalidPlaylist] = useState<boolean>(false);
 
   const [isUserAccountModalOpen, setUserAccountModalState] =
     useState<boolean>(false);
@@ -60,7 +56,11 @@ const CustomGame: React.FC<{ apiOrigin: string }> = ({ apiOrigin }) => {
     setUserAccountModalState(false);
   };
 
-  const location = useLocation();
+  useEffect(() => {
+    fetchData();
+    fetchAllSongs();
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, []);
 
   useEffect(() => {
     if (localStorage.getItem("firstTimeUser") !== "false") {
@@ -69,68 +69,51 @@ const CustomGame: React.FC<{ apiOrigin: string }> = ({ apiOrigin }) => {
     }
   }, []);
 
-  useEffect(() => {
-    const queryParams = queryString.parse(location.search);
-    const playlistId = queryString.parse(location.search).playlist;
+  const fetchData = () => {
+    const timezone = Intl.DateTimeFormat().resolvedOptions().timeZone;
+    fetch(`${apiOrigin}/api/dailySong?timeZone=${timezone}`)
+      .then((response) => response.json())
+      .then((data) => {
+        const userData = JSON.parse(
+          localStorage.getItem("userData") || "null"
+        ) || { main: [], custom: {} };
+        if (!localStorage.getItem("userData")) {
+          localStorage.setItem(
+            "userData",
+            JSON.stringify({ main: [], custom: {} })
+          );
+        }
 
-    if (playlistId) {
-      const timezone = Intl.DateTimeFormat().resolvedOptions().timeZone;
-      fetch(
-        `${apiOrigin}/api/playlist/${playlistId}/dailySong?timeZone=${timezone}${
-          queryParams.r ? "&r=1" : ""
-        }`,
-        { method: "GET" }
-      )
-        .then((response) => response.json())
-        .then((data) => {
-          const userData = JSON.parse(
-            localStorage.getItem("userData") || "null"
-          ) || { main: [], custom: {} };
-          if (!localStorage.getItem("userData")) {
-            localStorage.setItem(
-              "userData",
-              JSON.stringify({ main: [], custom: {} })
-            );
+        const isLastDataObjectMatchingId =
+          Array.isArray(userData.main) &&
+          userData.main.length > 0 &&
+          userData.main[userData.main.length - 1].id === data.id;
+
+        setSong(data.song);
+        setArtists(data.artists);
+        setId(data.id);
+        setTrackPreview(data.trackPreview);
+        setAlbumCover(data.albumCover);
+        setExternalUrl(data.externalUrl);
+
+        if (isLastDataObjectMatchingId) {
+          setUserGuesses(userData.main[userData.main.length - 1].guessList);
+          if (userData.main[userData.main.length - 1].hasFinished) {
+            setGameFinished(true);
           }
+        }
+      })
+      .catch((err) => console.error(err));
+  };
 
-          const customPlaylistObject = userData.custom[playlistId as string];
-          const isLastDataObjectMatchingId =
-            customPlaylistObject &&
-            customPlaylistObject[customPlaylistObject.length - 1] &&
-            customPlaylistObject[customPlaylistObject.length - 1].id ===
-              data.id;
-
-          setvalidPlaylist(true);
-          setSong(data.song);
-          setArtists(data.artists);
-          setId(data.id);
-          setTrackPreview(data.trackPreview);
-          setAlbumCover(data.albumCover);
-          setExternalUrl(data.externalUrl);
-
-          if (isLastDataObjectMatchingId) {
-            setUserGuesses(
-              customPlaylistObject[customPlaylistObject.length - 1].guessList
-            );
-            if (
-              customPlaylistObject[customPlaylistObject.length - 1].hasFinished
-            ) {
-              setGameFinished(true);
-            }
-          }
-
-          fetch(`${apiOrigin}/api/playlist/${playlistId}/allSongs`, {
-            method: "GET",
-          })
-            .then((response) => response.json())
-            .then((data) => {
-              setSongsInDb(data.tracklist);
-            })
-            .catch((err) => console.error(err));
-        })
-        .catch((err) => console.error(err));
-    }
-  }, [apiOrigin, location.search]);
+  const fetchAllSongs = () => {
+    fetch(`${apiOrigin}/api/allSongs`)
+      .then((response) => response.json())
+      .then((data) => {
+        setSongsInDb(data.tracklist);
+      })
+      .catch((err) => console.error(err));
+  };
 
   const handleUserGuessesUpdate = (newGuesses: TrackGuessFormat[]) => {
     setUserGuesses(newGuesses);
@@ -145,8 +128,7 @@ const CustomGame: React.FC<{ apiOrigin: string }> = ({ apiOrigin }) => {
     if (isGameFinished) {
       setGameFinished(true);
       const timezone = Intl.DateTimeFormat().resolvedOptions().timeZone;
-      const playlistId = queryString.parse(location.search).playlist;
-      fetch(`${apiOrigin}/api/playlist/${playlistId}/stats`, {
+      fetch(`${apiOrigin}/api/stats`, {
         method: "POST",
         headers: {
           "Content-Type": "application/json",
@@ -170,48 +152,35 @@ const CustomGame: React.FC<{ apiOrigin: string }> = ({ apiOrigin }) => {
       localStorage.getItem("userData") ?? '{ "main": [], "custom": {} }'
     );
 
-    const playlistId = queryString.parse(location.search).playlist;
-    const playlistData = data.custom[playlistId as string];
+    const isLastDataObjectMatchingId =
+      data.main.length > 0 && data.main[data.main.length - 1].id === id;
 
-    if (playlistData) {
-      const isLastDataObjectMatchingId =
-        playlistData.length > 0 &&
-        playlistData[playlistData.length - 1].id === id;
-
-      if (isLastDataObjectMatchingId) {
-        playlistData[playlistData.length - 1] = todaysDataObject;
-      } else {
-        playlistData.push(todaysDataObject);
-      }
-      data.custom[playlistId as string] = playlistData;
+    if (isLastDataObjectMatchingId) {
+      data.main[data.main.length - 1] = todaysDataObject;
     } else {
-      data.custom[playlistId as string] = [todaysDataObject];
+      data.main.push(todaysDataObject);
     }
 
     localStorage.setItem("userData", JSON.stringify(data));
   };
 
   useEffect(() => {
-    const queryParams = queryString.parse(location.search);
-    const playlistId = queryParams.playlist;
     const localData: SavedGameData = JSON.parse(
       localStorage.getItem("userData") ?? '{ "main": [], "custom": {} }'
     );
 
-    if (playlistId && localData.custom[playlistId as string]) {
-      const barHeights = calculateBarHeights(
-        localData.custom[playlistId as string]
-      );
+    if (localData) {
+      const barHeights = calculateBarHeights(localData.main);
       setStatsBarHeights(barHeights);
 
       const { statsNumCorrectString, statsCorrectPercentageString } =
-        calculateStatsBottom(localData.custom[playlistId as string]);
+        calculateStatsBottom(localData.main);
       setStatsCorrectString(statsNumCorrectString);
       setStatsCorrectPercentageString(statsCorrectPercentageString);
     } else {
       setStatsBarHeights(Array(7).fill(0));
     }
-  }, [isStatsModalOpen, location.search]);
+  }, [isStatsModalOpen]);
 
   return (
     <div className="font-sf-pro">
@@ -220,17 +189,12 @@ const CustomGame: React.FC<{ apiOrigin: string }> = ({ apiOrigin }) => {
         setStatsModal={setStatsModalState}
         setUAModel={setUserAccountModalState}
       />
-      {!trackPreview && self.location.search.split("&")[1] === "r=1" && (
+      {!trackPreview && (
         <div id="loader">
           <Loader></Loader>
         </div>
       )}
-      {!validPlaylist && (
-        <div id="playlist-search">
-          <PlaylistSearch></PlaylistSearch>
-        </div>
-      )}
-      {!gameFinished && trackPreview && validPlaylist && (
+      {!gameFinished && trackPreview && (
         <div id="game">
           <Game
             song={song}
@@ -250,7 +214,8 @@ const CustomGame: React.FC<{ apiOrigin: string }> = ({ apiOrigin }) => {
             userGuesses={userGuesses}
             id={id}
             albumCover={albumCover}
-            externalUrl={externalUrl}></GameConclusion>
+            externalUrl={externalUrl}
+          ></GameConclusion>
         </div>
       )}
 
@@ -260,33 +225,36 @@ const CustomGame: React.FC<{ apiOrigin: string }> = ({ apiOrigin }) => {
         onRequestClose={onRequestCloseHelpModal}
         className="bg-[#131213] text-white border-gray-800 border-2 p-10 mx-auto max-w-xs md:max-w-lg text-center"
         overlayClassName="overlay"
-        ariaHideApp={false}>
+        ariaHideApp={false}
+      >
         <HelpModal
-          onRequestCloseHelpModal={onRequestCloseHelpModal}></HelpModal>
+          onRequestCloseHelpModal={onRequestCloseHelpModal}
+        ></HelpModal>
       </Modal>
       <Modal
         isOpen={isStatsModalOpen}
         onRequestClose={closeStatsModal}
         className="bg-[#131213] text-white border-gray-800 border-2 p-10 mx-auto max-w-xs md:max-w-lg text-center"
         overlayClassName="overlay"
-        ariaHideApp={false}>
+        ariaHideApp={false}
+      >
         <StatsModal
           statsBarHeights={statsBarHeights}
           statsCorrectString={statsCorrectString}
-          statsCorrectPercentageString={
-            statsCorrectPercentageString
-          }></StatsModal>
+          statsCorrectPercentageString={statsCorrectPercentageString}
+        ></StatsModal>
       </Modal>
       <Modal
         isOpen={isUserAccountModalOpen}
         onRequestClose={closeUAModel}
         className="bg-[#131213] text-white border-gray-800 border-2 p-10 mx-auto max-w-xs md:max-w-lg text-center"
         overlayClassName="overlay"
-        ariaHideApp={false}>
+        ariaHideApp={false}
+      >
         <UserAccountModal apiOrigin={apiOrigin}></UserAccountModal>
       </Modal>
     </div>
   );
 };
 
-export default CustomGame;
+export default BaseGame;
