+++ src/client/src/types/GameResult.ts
