+++ src/client/src/types/SavedGameData.ts
