+++ src/client/tailwind.config.js
@@ -1,6 +1,6 @@
 /** @type {import('tailwindcss').Config} */
 export default {
-  content: ["./src/components/**/*.tsx"],
+  content: ["./src/**/*.tsx"],
   theme: {
     extend: {
       fontFamily: {
